Basic changes to admin portal css Moved main portal/site into a folder titled 'MainPortal' to increase navigatability
