+++ src/MainPortal/Components/Footer.js
@@ -1,5 +1,5 @@
 import React from "react";
-import '../App.css'; 
+import '../../SiteCss/App.css';
 
 function Footer(){
     return (
