+++ src/MainPortal/Pages/Contact.js
@@ -1,5 +1,5 @@
 import React from "react";
-import contactImage from '../assets/hero.jpeg';
+import contactImage from '../../assets/hero.jpeg';
 import '../PagesCSS/Contact.css';
 
 const Contact = () => {
