+++ src/SiteCss/AdminApp.css
@@ -7,8 +7,8 @@
 
 .AdminApp {
     font-family: "Trebuchet MS", "Lucida Sans Unicode", "Lucida Grande", "Lucida Sans", Arial, sans-serif;
-    background-color: #2a3447;
-    color: white;
+    background-color: #1d1f20;
+    color: #222020;
 }
 
 .AdminApp .container {
@@ -18,7 +18,7 @@
 .AdminApp .menuContainer {
     width: 250px;
     padding: 5px 20px;
-    border-right: 2px solid #384256;
+    border-right: 2px solid #f1f1f1;
 }
 
 .AdminApp .contentContainer {
